+++ pages/index.js
@@ -1,22 +1,21 @@
-import Head from "next/head";
-import Image from "next/image";
-import Layout from "../components/layout";
-import Hero from "../components/home/hero";
-import Animation from "../components/animation";
+import Head from 'next/head'
+import Layout from '../components/layout';
+import Hero from '../components/home/hero';
+import Animation from '../components/home/animation';
 
 export default function Home() {
   return (
     <Layout>
-      <Head>
-        <title>Sarah 포토폴리오</title>
-        <meta name="description" content="포토폴리오입니다" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-       <section className="flex min-h-screen flex-col items-center justify-center text-gray-600 body-font">
+        <Head>
+            <title>Sarah 포트폴리오</title>
+            <meta name="description" content="오늘도 화이팅" />
+            <link rel="icon" href="/favicon.ico" />
+        </Head>
+        <section className="flex min-h-screen flex-col items-center justify-center text-gray-600 body-font">
             <div className="container mx-auto flex px-5 py-24 md:flex-row flex-col items-center">
-          <Hero />
+                <Hero/>
             </div>
         </section>
-      </Layout>
-  )
+    </Layout>
+  );
 }
